cloud_functions/gemini_monitor/requirements.txt
functions-framework==3.8.1
google-cloud-bigquery==3.25.0
google-genai==1.16.1
